Create exercise_06_07.c
Thanks for doing this repo.  I currently reading through the book and have found it helpful to check my answers with yours.  Found a bug in this code that results in getting a lot of numbers listed as prime that are not prime.  The problem is you have j++ after the 2nd while loop which results in that 2nd while loop never being used again.  Your first time through the 2nd while loop i = 2 and the loop stops when j=75 as 2*75 = 150.  So the while loop ends and you increment i up to 3.  But you also increment j so you get j = 76.   The next time through the while loop you then have 3*76 which is greater than 150 so the while loop will never again be used.  9 for instance is listed as a prime number, but if you just change j=i after you increment i then you get the desired result.  But again big thanks for doing this repo.  Best, Kevin